% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spatial_graph.R
\name{build_spatial_graph}
\alias{build_spatial_graph}
\title{build_spatial_graph}
\usage{
build_spatial_graph(cell_data, cell_id = "cell_id", k = 6, max_dist = 100)
}
\arguments{
\item{max_dist}{}
}
\description{
build_spatial_graph
}
