% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaBarrierQ.R
\name{SpaBarrierQ}
\alias{SpaBarrierQ}
\title{SpaBarrierQ
-Blocking count: The number of stromal cells (excluding themselves) that are simultaneously located within the "radius around immune cells" and the "radius around nearest tumor cells".
Principle:
\itemize{
\item Take the nearest tumor cell as the anchor point and expand radius_factor × tumor distance outward as the search radius;
}}
\usage{
SpaBarrierQ(
  cell_data,
  type,
  tumor_type = "Tumor",
  immune_type = "Immune",
  stroma_type = "Stroma",
  radius_factor = 1.5
)
}
\arguments{
\item{cell_data}{data.frame; It must contain four columns: x, y, cell_id and type}

\item{tumor_type}{Anchor cell type, default "Tumor"}

\item{immune_type}{Blocked cell type, default "Immune"}

\item{stroma_type}{Barrier cell type, default "Stroma"}

\item{radius_factor}{search radius multiple (default 1.5), ≥1}
}
\value{
The original table adds the column stroma_blocker (integer ≥ 0)
}
\description{
SpaBarrierQ
-Blocking count: The number of stromal cells (excluding themselves) that are simultaneously located within the "radius around immune cells" and the "radius around nearest tumor cells".
Principle:
\itemize{
\item Take the nearest tumor cell as the anchor point and expand radius_factor × tumor distance outward as the search radius;
}
}
\examples{
cell_data <- SpaBarrierQ(cell_data,type,tumor_type = "Tumor",mmune_type = "Immune",stroma_type = "Stroma")
}
