% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaComLeiden.R
\name{SpaComLeiden}
\alias{SpaComLeiden}
\title{SpaComLeiden}
\usage{
SpaComLeiden(cell_data, edges, resolution = 0.5, n_iter = 10)
}
\arguments{
\item{cell_data}{data.frame；It must contain four columns: x, y, cell_id and type}

\item{edges}{data.frame；It must contain three columns: from, to, and weight}

\item{resolution}{Resolution parameter (default 0.5}

\item{n_iter}{Number of iterations (default: 10}
}
\value{
The igraph object has added the properties of community, degree, betweenness and closeness
}
\description{
SpaComLeiden
}
\examples{
g <- SpaComLeiden(cell_data, edges, resolution = 0.5, n_iter = 10)
}
