% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaDensityMk.R
\name{SpaDensityMk}
\alias{SpaDensityMk}
\title{SpaDensityMk
Core Logic:
\enumerate{
\item Count of neighbors within the radius that do not include itself;
\item Divide by the area of the circle (mm²) to obtain the physical density;
}}
\usage{
SpaDensityMk(cell_data, radius = 50)
}
\arguments{
\item{cell_data}{data.frame; It must contain x, y (unit: µm) and any other column}

\item{radius}{Circular window Radius (µm, default 50)}
}
\value{
The original table adds a new column density_mm2 (cells/mm²)
}
\description{
SpaDensityMk
Core Logic:
\enumerate{
\item Count of neighbors within the radius that do not include itself;
\item Divide by the area of the circle (mm²) to obtain the physical density;
}
}
\examples{
cell_data <- SpaDensityMk(cell_data, radius = 20)
}
