% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaGeoNet.R
\name{SpaGeoNet}
\alias{SpaGeoNet}
\title{SpaGeoNet
Construct Spatial Geometry Network (Spatial Geometry Network).
Generate an undirected, weighted adjacency edge list based on the KNN + distance threshold
Core Logic:
\enumerate{
\item Take k nearest neighbors (excluding the cell itself) for each cell;
}}
\usage{
SpaGeoNet(cell_data, cell_id = "cell_id", k = 50, max_dist = 100)
}
\arguments{
\item{cell_data}{data.frame; It must contain four columns: x, y, cell_id, and type (unit: µm).}

\item{cell_id}{Cell Unique ID column name (default "cell_id")}

\item{k}{Number of nearest neighbors (default 50, excluding itself)}

\item{max_dist}{Maximum side length (µm, default 100)}
}
\value{
data.frame: from, to, weight, type four columns, directly graph_from_data_frame()
}
\description{
SpaGeoNet
Construct Spatial Geometry Network (Spatial Geometry Network).
Generate an undirected, weighted adjacency edge list based on the KNN + distance threshold
Core Logic:
\enumerate{
\item Take k nearest neighbors (excluding the cell itself) for each cell;
}
}
\examples{
edges <- SpaGeoNet(cell_data, cell_id='cell_id', k = 6, max_dist = 50)
}
