% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaNetPlot.R
\name{SpaNetPlot}
\alias{SpaNetPlot}
\title{SpaNetPlot Draw the spatial community network}
\usage{
SpaNetPlot(g, node_alpha = 0.7, title = "Spatial Community Clsuter")
}
\arguments{
\item{g}{igraph object; It must contain attributes such as x, y, name, community, and degree}

\item{node_alpha}{node transparency (default 0.7)}

\item{title}{Graph Title (default: "Spatial Community Cluster")}
}
\value{
ggplot object
}
\description{
SpaNetPlot Draw the spatial community network
}
\examples{
SpaNetPlot(g, node_alpha = 0.7)
}
