% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaProxTarget.R
\name{SpaProxTarget}
\alias{SpaProxTarget}
\title{SpaProxTarget
Based on the Euclidean distance, excluding itself, return:
-dist_to_nearest: The straight-line distance (µm) to the nearest target cell
-count_in_radius: Count of target cells within the radius (excluding itself)}
\usage{
SpaProxTarget(cell_data, type, target_type = "Tumor", radius = 50)
}
\arguments{
\item{cell_data}{data.frame; It must contain four columns: x, y, cell_id and type}

\item{type}{Cell type column name (default "type")}

\item{target_type}{Target cell type, such as "Tumor" (default)}

\item{radius}{Circular search radius (µm, default 50)}
}
\value{
Two new columns have been added to the original table: dist_to_nearest and count_in_radius
}
\description{
SpaProxTarget
Based on the Euclidean distance, excluding itself, return:
-dist_to_nearest: The straight-line distance (µm) to the nearest target cell
-count_in_radius: Count of target cells within the radius (excluding itself)
}
\examples{
target_cells <- SpaProxTarget(cell_data, type, "Tumor", 50)
}
